refactor(agent-loop): decompose capability stage mechanics
